Add files in respective folders
