+++ C++/Algorithms/String Algorithms/pattern searching (KMP).cpp
