+++ C++/Data Structures/Trees/leftviewofabinarytree.cpp
