+++ C++/Extra questions/Magic_square_generator.cpp
