+++ C++/Extra questions/Pascal's Triangle
