+++ C++/Extra questions/PlaceNQueen.cpp
