+++ C++/Extra questions/countandsay.cpp
