+++ C++/Extra questions/gas_station.cpp
