+++ C++/Extra questions/twinPrime.cpp
