+++ C++/Extra questions/vectorsort.cpp
