+++ JAVASCRIPT/binary_search.js
@@ -1,31 +1,37 @@
-// binary search code in javascript
+<script>
+let recursiveFunction = function (arr, x, start, end) {
+	
+	// Base Condition
+	if (start > end) return false;
 
-function binarySearch(sortedArray, key) {
-    let start = 0;
-    let end = sortedArray.length - 1;
+	// Find the middle index
+	let mid=Math.floor((start + end)/2);
 
-    while (start <= end) {
-        let middle = Math.floor((start + end) / 2);
+	// Compare mid with given key x
+	if (arr[mid]===x) return true;
+		
+	// If element at mid is greater than x,
+	// search in the left half of mid
+	if(arr[mid] > x)
+		return recursiveFunction(arr, x, start, mid-1);
+	else
 
-        if (sortedArray[middle] === key) {
-            // found the key
-            return middle;
-        } else if (sortedArray[middle] < key) {
-            // continue searching to the right
-            start = middle + 1;
-        } else {
-            // search searching to the left
-            end = middle - 1;
-        }
-    }
-    // key wasn't found
-    return -1;
+		// If element at mid is smaller than x,
+		// search in the right half of mid
+		return recursiveFunction(arr, x, mid+1, end);
 }
 
-// array (array should be sorted)
-let array = [11, 22, 33, 44, 55, 66, 77, 88, 99]
+// Driver code
+let arr = [1, 3, 5, 7, 8, 9];
+let x = 5;
 
-// key to be searched in array
-let key = 44;
+if (recursiveFunction(arr, x, 0, arr.length-1))
+	document.write("Element found!<br>");
+else document.write("Element not found!<br>");
 
-console.log(binarySearch(array, key))
+x = 6;
+
+if (recursiveFunction(arr, x, 0, arr.length-1))
+	document.write("Element found!<br>");
+else document.write("Element not found!<br>");
+</script>
